requirements.txt
# Core Data Science (STABLE VERSIONS - Windows compatible)
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
imbalanced-learn==0.11.0
shap==0.42.1
joblib==1.3.2
scipy==1.10.1

# Backend API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Google AI
google-genai>=1.73.0

# WhatsApp Integration (optional - add when needed)
twilio==8.10.0

# Frontend Dashboard (optional - add when needed)
streamlit==1.28.1
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.0

# Code Quality (optional - for advanced users)
black==23.12.1
flake8==6.1.0
mypy==1.8.0